mypy for resources

+++ sky/clouds/azure.py
@@ -147,7 +147,7 @@ class Azure(clouds.Cloud):
 
     @classmethod
     def regions_with_offering(cls, instance_type: Optional[str],
-                              accelerators: Optional[Dict[str, int]],
+                              accelerators: Optional[Dict[str, float]],
                               use_spot: bool, region: Optional[str],
                               zone: Optional[str]) -> List[clouds.Region]:
         del accelerators  # unused
@@ -171,7 +171,7 @@ class Azure(clouds.Cloud):
         cls,
         *,
         instance_type: Optional[str] = None,
-        accelerators: Optional[Dict[str, int]] = None,
+        accelerators: Optional[Dict[str, float]] = None,
         use_spot: bool = False,
     ) -> Iterator[Tuple[clouds.Region, List[clouds.Zone]]]:
         regions = cls.regions_with_offering(instance_type,
@@ -335,7 +335,7 @@ class Azure(clouds.Cloud):
 
     def accelerator_in_region_or_zone(self,
                                       accelerator: str,
-                                      acc_count: int,
+                                      acc_count: float,
                                       region: Optional[str] = None,
                                       zone: Optional[str] = None) -> bool:
         return service_catalog.accelerator_in_region_or_zone(
