+++ sky/clouds/local.py
@@ -47,7 +47,7 @@ class Local(clouds.Cloud):
         cls,
         *,
         instance_type: Optional[str] = None,
-        accelerators: Optional[Dict[str, int]] = None,
+        accelerators: Optional[Dict[str, float]] = None,
         use_spot: bool = False,
     ) -> Iterator[Tuple[clouds.Region, List[clouds.Zone]]]:
         del instance_type
@@ -124,7 +124,7 @@ class Local(clouds.Cloud):
 
     def accelerator_in_region_or_zone(self,
                                       accelerator: str,
-                                      acc_count: int,
+                                      acc_count: float,
                                       region: Optional[str] = None,
                                       zone: Optional[str] = None) -> bool:
         # In the public cloud case, an accelerator may not be in a region/zone.
