+++ sky/clouds/service_catalog/aws_catalog.py
@@ -75,7 +75,7 @@ def validate_region_zone(
 
 
 def accelerator_in_region_or_zone(acc_name: str,
-                                  acc_count: int,
+                                  acc_count: float,
                                   region: Optional[str] = None,
                                   zone: Optional[str] = None) -> bool:
     return common.accelerator_in_region_or_zone_impl(_df, acc_name, acc_count,
